# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Policy Platform API",
    "description": "Enterprise policy ingestion and immutable versioning service. Provides ingestion batches, blob registration, queue-driven extraction, and audit-friendly traceability.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/ingest/batches": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Create ingestion batch",
        "description": "Creates an ingestion batch to group uploaded policy documents.",
        "operationId": "create_batch_v1_ingest_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionBatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batches/{batch_id}": {
      "get": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Get ingestion batch",
        "description": "Fetches an ingestion batch (tenant-scoped).",
        "operationId": "get_batch_v1_ingest_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Batch UUID",
              "title": "Batch Id"
            },
            "description": "Batch UUID"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batches/{batch_id}/upload-urls": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Generate upload SAS URL",
        "description": "Generates a time-limited SAS URL for client upload to Azure Blob Storage.",
        "operationId": "generate_upload_url_v1_ingest_batches__batch_id__upload_urls_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Batch UUID",
              "title": "Batch Id"
            },
            "description": "Batch UUID"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batches/{batch_id}/register": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Register uploaded policy document",
        "description": "Registers a previously-uploaded blob as an ingestion item and creates an immutable policy version. Also enqueues background extraction via Azure Storage Queue.",
        "operationId": "register_uploaded_document_v1_ingest_batches__batch_id__register_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Batch UUID",
              "title": "Batch Id"
            },
            "description": "Batch UUID"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestionBatchCreateRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "submitted_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By User Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "IngestionBatchCreateRequest"
      },
      "IngestionBatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "submitted_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "source_system",
          "submitted_by_user_id",
          "status",
          "status_reason",
          "correlation_id",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionBatchResponse"
      },
...
      "RegisterDocumentRequest": {
        "properties": {
          "container_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Container Name"
          },
          "blob_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Blob Path"
          },
          "policy_external_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Policy External Id"
          },
          "policy_name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Policy Name"
          },
          "version_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Label",
            "description": "Optional client version label (unique per policy)"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "submitted_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By User Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "blob_version_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Version Id"
          },
          "blob_etag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Etag"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "content_length": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Length"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "container_name",
          "blob_path",
          "policy_external_id",
          "policy_name"
        ],
        "title": "RegisterDocumentRequest"
      },
      "RegisterDocumentResponse": {
        "properties": {
          "ingest_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ingest Item Id"
          },
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id"
          },
          "policy_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Version Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "content_sha256": {
            "type": "string",
            "title": "Content Sha256"
          },
          "metadata_sha256": {
            "type": "string",
            "title": "Metadata Sha256"
          },
          "parse_status": {
            "type": "string",
            "title": "Parse Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ingest_item_id",
          "policy_id",
          "policy_version_id",
          "version_number",
          "content_sha256",
          "metadata_sha256",
          "parse_status"
        ],
        "title": "RegisterDocumentResponse"
      },
      "UploadUrlRequest": {
        "properties": {
          "container_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Container Name",
            "description": "Azure Blob container"
          },
          "blob_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Blob Path",
            "description": "Blob path/name"
          },
          "expires_in_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Minutes"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "container_name",
          "blob_path"
        ],
        "title": "UploadUrlRequest"
      },
      "UploadUrlResponse": {
        "properties": {
          "upload_sas_url": {
            "type": "string",
            "title": "Upload Sas Url"
          },
          "blob_uri": {
            "type": "string",
            "title": "Blob Uri"
          },
          "expires_in_minutes": {
            "type": "integer",
            "title": "Expires In Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_sas_url",
          "blob_uri",
          "expires_in_minutes"
        ],
        "title": "UploadUrlResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}